Add Database empty script

+++ DigitalHubLMS.Core/Data/Entities/ClassQuizAnswer.cs
@@ -22,9 +22,9 @@ namespace DigitalHubLMS.Core.Data.Entities
         [Column("option_id")]
         public long OptionId { get; set; }
         [Column("attempt")]
-        public byte Attempt { get; set; }
+        public short Attempt { get; set; }
         [Column("score")]
-        public byte Score { get; set; }
+        public short Score { get; set; }
         [Column("created_at")]
         public DateTime? CreatedAt { get; set; }
         [Column("updated_at")]
